# Shader processor include optimization