I have a pretty big scene (100+ meshed, 50+ shaders) in glb format, when it loaded the loop in the `ShaderProcessor._ProcessIncludes()` was called 4927 times, and takes 200.57 ms and each loop take 0.0407ms. This values bases on 30 runs.
After performance optimzations I got next result — the loop calls 3955 times, takes 70.82ms for all calls and 0.0179ms per one loop.

1. I make regex as a static fields to avoid builing it in runtime.
2. Change string replaces from regex to regular string where it possible.
3. Replace while loop to for-of to avoid uneccessary `.exec()` at end of while-loop.
4. Replace very heavy string replace opertation to `parts.join("")` (more then 2x boost here)
5. Fix warning from `new String()`.

Every change puts in a single commit.

For future discussing - we also can change this part:
```ts
for (let index = 0; index < splits.length; index += 2) {
    const source = new RegExp(splits[index], "g");
    const dest = splits[index + 1];
    includeContent = includeContent.replace(source, dest);
}
```

with `.replaceAll` and save 5-7ms for all loading time (it would be 65ms for all calls in my case). But now I can't use it because `TS2550: Property 'replaceAll' does not exist on type 'string'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2021' or later.`