@Popov72 2023-06-04
The build is failing with the error:

> generated/Engines/Processors/shaderProcessor.ts(383,29): error TS2802: Type 'IterableIterator<RegExpMatchArray>' can only be iterated through when using the '--downlevelIteration' flag or with a '--target' of 'es2015' or higher.

cc @RaananW in case we can make something about it, but you may have to use iterators "by hand" instead of the `for ... of` construct.

Also, `TS2550: Property 'replaceAll' does not exist on type 'string'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2021' or later.` is very strange, are you sure you are not passing it wrong parameters?