@RaananW 2023-06-05
> cc @RaananW in case we can make something about it, but you may have to use iterators "by hand" instead of the `for ... of` construct.

Our UMD version is build with an old(er) target due to backwards compatibility. We will need to see if anyone is influenced by moving the target to es2015. But yes -  using a different style of iteration (like we use throughout the framework) would be a better solution.




> TS2550: Property 'replaceAll' does not exist on type 'string'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2021' or later

I don't mind adding "ES2021.String" as a lib to support replaceAll (typescript is right - it was added in es2021) you could also add the declaration to our mixins/polyfill file, which will have the same impact

```
interface String {
    /**
     * Replace all instances of a substring in a string, using a regular expression or search string.
     * @param searchValue A string to search for.
     * @param replaceValue A string containing the text to replace for every successful match of searchValue in this string.
     */
    replaceAll(searchValue: string | RegExp, replaceValue: string): string;

    /**
     * Replace all instances of a substring in a string, using a regular expression or search string.
     * @param searchValue A string to search for.
     * @param replacer A function that returns the replacement text.
     */
    replaceAll(searchValue: string | RegExp, replacer: (substring: string, ...args: any[]) => string): string;
}
```