@RaananW 2023-06-08
> @Popov72 @RaananW there is an error in CI:
> ```
> Failed tasks:
>    
>    - @lts/core:build
> 
> ##[warning]Couldn't find a debug log in the cache or working directory
> ##[error]Error: Npm failed with return code: 1
> ```
> 
> Do you know how to fix it?

the error i see is the iterator (es2015) during the umd build. Where do you see the npm issue?